Move new schedulers into main crate
